Referral api with tests

+++ NotificationUtil/SMS/Repository/TextLocalResponse.cs
@@ -14,36 +14,36 @@ namespace NotificationUtil.Mode.SMS;
 
 public class Message
 {
-    public int num_parts { get; set; }
+    public long num_parts { get; set; }
     public string sender { get; set; }
     public string content { get; set; }
 }
 
 public class ReceiverInfo
 {
-    public int id { get; set; }
+    public long id { get; set; }
     public long recipient { get; set; }
 }
 
 public class Warning
 {
-    public int code { get; set; }
+    public long code { get; set; }
     public string message { get; set; }
 }
 
 public class Error
 {
-    public int code { get; set; }
+    public long code { get; set; }
     public string message { get; set; }
 }
 
 public class TextLocalResponse
 {
     public bool test_mode { get; set; }
-    public int balance { get; set; }
-    public int batch_id { get; set; }
-    public int cost { get; set; }
-    public int num_messages { get; set; }
+    public long balance { get; set; }
+    public long batch_id { get; set; }
+    public long cost { get; set; }
+    public long num_messages { get; set; }
     public Message message { get; set; }
     public string receipt_url { get; set; }
     public string custom { get; set; }
